Feat: Move to separated strurcture
