+++ src/AthenaTwitchBot/models/bot/contexts/__init__.py
@@ -7,12 +7,7 @@ from __future__ import annotations
 # Custom Library
 
 # Custom Packages
-from AthenaTwitchBot.models.data_handlers.data_handler import DataHandler
 
 # ----------------------------------------------------------------------------------------------------------------------
 # - Code -
 # ----------------------------------------------------------------------------------------------------------------------
-class DataHandlerTwitch(DataHandler):
-    def handle(self, data:bytearray):
-        print(data)
-        pass
