+++ src/AthenaTwitchBot/models/bot/data_handlers/data_handler.py
@@ -3,15 +3,17 @@
 # ----------------------------------------------------------------------------------------------------------------------
 # General Packages
 from __future__ import annotations
+from abc import ABC, abstractmethod
 
 # Custom Library
 
 # Custom Packages
-from AthenaTwitchBot.models.outputs.output import Output
+from AthenaTwitchBot.models.bot.contexts.context import Context
 
 # ----------------------------------------------------------------------------------------------------------------------
 # - Code -
 # ----------------------------------------------------------------------------------------------------------------------
-class OutputConsole(Output):
-    def undefined(self, data):
-        pass
+class DataHandler(ABC):
+    @abstractmethod
+    def handle(self, data:bytearray) -> Context:
+        """handles the data in a proper manner"""
