+++ src/AthenaTwitchBot/models/bot/data_handlers/data_handler_twitch.py
@@ -1,0 +1,18 @@
+# ----------------------------------------------------------------------------------------------------------------------
+# - Package Imports -
+# ----------------------------------------------------------------------------------------------------------------------
+# General Packages
+from __future__ import annotations
+
+# Custom Library
+
+# Custom Packages
+from AthenaTwitchBot.models.bot.data_handlers.data_handler import DataHandler
+
+# ----------------------------------------------------------------------------------------------------------------------
+# - Code -
+# ----------------------------------------------------------------------------------------------------------------------
+class DataHandlerTwitch(DataHandler):
+    def handle(self, data:bytearray):
+        print(data)
+        pass
