+++ src/AthenaTwitchBot/models/bot/outputs/output.py
