+++ src/AthenaTwitchBot/models/bot/protocols/protocol_server.py
@@ -8,8 +8,8 @@ from dataclasses import dataclass
 # Custom Library
 
 # Custom Packages
-from AthenaTwitchBot.models.protocols.protocol import Protocol
-from AthenaTwitchBot.models.data_handlers.data_handler_server import DataHandlerServer
+from AthenaTwitchBot.models.bot.protocols.protocol import Protocol
+from AthenaTwitchBot.models.bot.data_handlers.data_handler_server import DataHandlerServer
 
 # ----------------------------------------------------------------------------------------------------------------------
 # - All -
